% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{DdiPpiCloNoGenericV5-package}
\alias{DdiPpiCloNoGenericV5}
\alias{DdiPpiCloNoGenericV5-package}
\title{DdiPpiCloNoGenericV5: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
